+++ src/server/devices/velux/index.ts
@@ -1,0 +1,4 @@
+export { VeluxDevice } from './veluxDevice';
+export { VeluxShutter } from './veluxShutter';
+export { VeluxDeviceRegistrationInfo } from './veluxDeviceRegistrationInfo';
+export { VeluxService } from './veluxService';