+++ src/server/devices/velux/veluxDevice.ts
@@ -1,0 +1,55 @@
+import { IoBrokerBaseDevice } from '../IoBrokerBaseDevice';
+import { iDisposable } from '../../services';
+import { IoBrokerDeviceInfo } from '../IoBrokerDeviceInfo';
+import { DeviceType } from '../deviceType';
+import { DeviceCapability } from '../DeviceCapability';
+import { LogLevel } from '../../../models';
+
+export class VeluxDevice extends IoBrokerBaseDevice implements iDisposable {
+  private _lastUpdate: Date = new Date(0);
+
+  public constructor(pInfo: IoBrokerDeviceInfo, pType: DeviceType) {
+    super(pInfo, pType);
+    this.deviceCapabilities.push(DeviceCapability.batteryDriven);
+  }
+
+  public get lastUpdate(): Date {
+    return this._lastUpdate;
+  }
+
+  /** @inheritDoc */
+  public update(idSplit: string[], state: ioBroker.State, initial: boolean = false, pOverride: boolean = false): void {
+    this.log(
+      LogLevel.DeepTrace,
+      `Smartgarden: ${initial ? 'Initiales ' : ''}Update: ID: ${idSplit.join('.')} JSON: ${JSON.stringify(state)}`,
+    );
+    if (!pOverride) {
+      this.log(
+        LogLevel.Warn,
+        `Keine Update Überschreibung:\n\tID: ${idSplit.join('.')}\n\tData: ${JSON.stringify(state)}`,
+      );
+    }
+
+    const stateLastUpdate: Date = new Date(state.ts);
+    if (stateLastUpdate > this._lastUpdate) {
+      this._lastUpdate = stateLastUpdate;
+    }
+    if (idSplit.length < 5) {
+      return;
+    }
+    const stateName: string = idSplit[4];
+    this.stateMap.set(stateName, state);
+    const individualCallbacks: Array<(val: ioBroker.StateValue) => void> | undefined =
+      this.individualStateCallbacks.get(stateName);
+    if (individualCallbacks !== undefined) {
+      for (const cb of individualCallbacks) {
+        cb(state.val);
+      }
+    }
+  }
+
+  /** @inheritDoc */
+  public dispose(): void {
+    // Nothing yet
+  }
+}