+++ src/server/devices/velux/veluxDeviceRegistrationInfo.ts
@@ -1,0 +1,11 @@
+import { DeviceType } from '../deviceType';
+
+export class VeluxDeviceRegistrationInfo {
+  public constructor(
+    public readonly deviceType: DeviceType,
+    public readonly deviceId: string,
+    public readonly deviceName: string,
+    public readonly room: string,
+    public readonly roomIndex: number,
+  ) {}
+}